Separate markup inference from selector inference
