+++ src/contentScript/nativeEditor/frameworks.ts
@@ -18,7 +18,7 @@
 import { Framework } from "@/messaging/constants";
 import adapters from "@/frameworks/adapters";
 import { isEmpty, uniq } from "lodash";
-import { inferSelectorsIncludingStableAncestors } from "@/contentScript/nativeEditor/infer";
+import { inferSelectorsIncludingStableAncestors } from "@/contentScript/nativeEditor/selectorInference";
 import { ElementInfo } from "@/contentScript/nativeEditor/types";
 
 export async function elementInfo(
