+++ src/contentScript/nativeEditor/markupInference.ts
@@ -15,64 +15,33 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { uniq, sortBy, unary, intersection, compact } from "lodash";
-import { getCssSelector } from "css-selector-generator";
-import { isNullOrBlank, mostCommonElement, matchesAnyPattern } from "@/utils";
-import { BusinessError } from "@/errors";
-import { CssSelectorType } from "css-selector-generator/types/types";
-import { $safeFind } from "@/helpers";
+import { isNullOrBlank, matchesAnyPattern, mostCommonElement } from "@/utils";
 import {
   EXTENSION_POINT_DATA_ATTR,
   PANEL_FRAME_ID,
   PIXIEBRIX_DATA_ATTR,
   PIXIEBRIX_READY_ATTRIBUTE,
 } from "@/common";
-import { guessUsefulness } from "@/utils/detectRandomString";
-import { getAttributeSelector } from "@/utils/selectorGenerator";
+import {
+  BUTTON_TAGS,
+  UNIQUE_ATTRIBUTES,
+} from "@/contentScript/nativeEditor/selectorInference";
+import { intersection, unary, uniq } from "lodash";
+import { BusinessError } from "@/errors";
 
-const BUTTON_TAGS: string[] = ["li", "button", "a", "span", "input", "svg"];
 const BUTTON_SELECTORS: string[] = ["[role='button']"];
 const ICON_TAGS = ["svg", "img"];
-const MENU_TAGS = ["ul", "tbody"];
 const CAPTION_TAGS = ["td", "a", "li", "span"];
 const MULTI_ATTRS = ["class", "rel"];
 const HEADER_TAGS = ["header", "h1", "h2", "h3", "h4", "h5", "h6"];
-// Layout tags that should be preserved during panel inference
+
+/** Layout tags that should be preserved during panel inference */
 const LAYOUT_TAGS = ["section", "header", "div", "article", "aside"];
 const TEXT_TAGS = ["span", "p", "b", "h1", "h2", "h3", "h4", "h5", "h6"];
-
 const ATTR_SKIP_ELEMENT_PATTERNS = [
   /^chevron-down$/,
   /^([\dA-Za-z]+)-chevron-down$/,
 ];
-
-export const UNIQUE_ATTRIBUTES: string[] = [
-  "id",
-  "name",
-
-  // Testing attributes
-  "data-cy", // Cypress
-  "data-testid",
-  "data-id",
-  "data-test",
-  "data-test-id",
-];
-// eslint-disable-next-line security/detect-non-literal-regexp -- Not user-provided
-const UNIQUE_ATTRIBUTES_REGEX = new RegExp(
-  UNIQUE_ATTRIBUTES.map((attribute) => `^\\[${attribute}=`).join("|")
-);
-
-export const UNIQUE_ATTRIBUTES_SELECTOR = UNIQUE_ATTRIBUTES.map(
-  (attribute) => `[${attribute}]`
-).join(",");
-
-function getUniqueAttributeSelectors(element: HTMLElement): string[] {
-  return compact(
-    UNIQUE_ATTRIBUTES.map((attribute) =>
-      getAttributeSelector(attribute, element.getAttribute(attribute))
-    )
-  );
-}
 
 /**
  * Attribute names to exclude from button/panel template inference.
@@ -99,7 +68,6 @@ const TEMPLATE_ATTR_EXCLUDE_PATTERNS = [
   // Exclude non-role aria attributes because they're generally unique across elements
   /^aria-(?!role).*$/,
 ];
-
 const TEMPLATE_VALUE_EXCLUDE_PATTERNS = new Map<string, RegExp[]>([
   ["class", [/^ember-view$/]],
   // eslint-disable-next-line security/detect-non-literal-regexp -- Our variables
@@ -108,41 +76,6 @@ const TEMPLATE_VALUE_EXCLUDE_PATTERNS = new Map<string, RegExp[]>([
 
 class SkipElement extends Error {
   override name = "SkipElement";
-}
-
-/** ID selectors and certain other attributes can uniquely identify items */
-function isSelectorUsuallyUnique(selector: string): boolean {
-  return selector.startsWith("#") || UNIQUE_ATTRIBUTES_REGEX.test(selector);
-}
-
-/**
- * Prefers unique selectors and classes. A lower number means higher preference. To be used with lodash.sortBy
- * @example
- * 0   '#best-link-on-the-page'
- * 1   '[data-cy="b4da55"]'
- * 2   '.navItem'
- * 2   '.birdsArentReal'
- * 21  '#name > :nth-child(2)'
- * 30  '[aria-label="Click elsewhere"]'
- */
-export function getSelectorPreference(selector: string): number {
-  if (selector.includes(":nth-child")) {
-    return selector.length;
-  }
-
-  if (selector.startsWith("#")) {
-    return 0;
-  }
-
-  if (isSelectorUsuallyUnique(selector)) {
-    return 1;
-  }
-
-  if (selector.startsWith(".")) {
-    return 2;
-  }
-
-  return selector.length;
 }
 
 function outerHTML(element: Element | string): string {
@@ -157,11 +90,6 @@ function newElement(tagName: string): Element {
   return document.createElement(tagName);
 }
 
-function escapeDoubleQuotes(str: string): string {
-  // https://gist.github.com/getify/3667624
-  return str.replace(/\\([\S\s])|(")/g, "\\$1$2");
-}
-
 /**
  * Returns true iff any of the immediate children are text nodes.
  * @param $element
@@ -170,7 +98,7 @@ function hasTextNodeChild(element: Element): boolean {
   return [...element.childNodes].some((x) => x.nodeType === Node.TEXT_NODE);
 }
 
-function commonAttribute(items: Element[], attribute: string) {
+export function commonAttribute(items: Element[], attribute: string) {
   const attributeValues = items.map(
     (x) => x.attributes.getNamedItem(attribute)?.value
   );
@@ -535,284 +463,6 @@ function commonPanelHTML(tag: string, $items: JQuery): string {
   return outerHTML(common);
 }
 
-const DEFAULT_SELECTOR_PRIORITIES: Array<keyof typeof CssSelectorType> = [
-  "id",
-  "tag",
-  "class",
-  "attribute",
-  "nthoftype",
-  "nthchild",
-];
-
-interface SafeCssSelectorOptions {
-  selectors?: Array<keyof typeof CssSelectorType>;
-  root?: Element;
-  excludeRandomClasses?: boolean;
-}
-
-/**
- * Calls getCssSelector with smarter handling of undefined root element and blacklisting common
- * front-end framework elements that aren't good for selectors
- */
-export function safeCssSelector(
-  element: HTMLElement,
-  {
-    selectors = DEFAULT_SELECTOR_PRIORITIES,
-    excludeRandomClasses = false,
-    // eslint-disable-next-line unicorn/no-useless-undefined -- Convert null to undefined or else getCssSelector bails
-    root = undefined,
-  }: SafeCssSelectorOptions = {}
-): string {
-  // https://github.com/fczbkk/css-selector-generator
-
-  const selector = getCssSelector(element, {
-    // NOTE: if you add an id pattern to the blacklist, you also need to add it to
-    // selectorGenerator.ts:getAttributeSelector
-    blacklist: [
-      // Emberjs component tracking
-      "#ember*",
-      // Salesforce Aura component tracking
-      "[data-aura-rendered-by]",
-      // Vuejs component tracking
-      "[data-v-*]",
-      // Our attributes
-      `[${EXTENSION_POINT_DATA_ATTR}='*']`,
-      `[${PIXIEBRIX_DATA_ATTR}='*']`,
-      `[${PIXIEBRIX_READY_ATTRIBUTE}='*']`,
-      excludeRandomClasses
-        ? (selector) => {
-            if (!selector.startsWith(".")) {
-              return false;
-            }
-
-            const usefulness = guessUsefulness(selector);
-            console.debug("css-selector-generator:  ", usefulness);
-            return usefulness.isRandom;
-          }
-        : undefined,
-    ],
-    whitelist: [
-      // Data attributes people use in automated tests are unlikely to change frequently
-      "[data-cy='*']",
-      "[data-testid='*']",
-      "[data-test='*']",
-    ],
-    selectors,
-    combineWithinSelector: true,
-    combineBetweenSelectors: true,
-    root,
-  });
-
-  if (root == null && selector.startsWith(":nth-child")) {
-    // JQuery will happily return other matches that match the nth-child chain, so make attach it to the body
-    // to get the expected CSS selector behavior
-    return `body${selector}`;
-  }
-
-  return selector;
-}
-
-function findAncestorsWithIdLikeSelectors(
-  element: HTMLElement,
-  root?: Element
-): HTMLElement[] {
-  // eslint-disable-next-line unicorn/no-array-callback-reference -- jQuery false positive
-  return $(element).parentsUntil(root).filter(UNIQUE_ATTRIBUTES_SELECTOR).get();
-}
-
-export function inferSelectorsIncludingStableAncestors(
-  element: HTMLElement,
-  root?: Element,
-  excludeRandomClasses?: boolean
-): string[] {
-  const stableAncestors = findAncestorsWithIdLikeSelectors(
-    element,
-    root
-  ).flatMap((stableAncestor) =>
-    inferSelectors(element, stableAncestor, excludeRandomClasses).flatMap(
-      (selector) =>
-        getUniqueAttributeSelectors(stableAncestor)
-          .filter(Boolean)
-          .map((stableAttributeSelector) =>
-            [stableAttributeSelector, selector].join(" ")
-          )
-    )
-  );
-
-  return sortBy(
-    compact(
-      uniq([
-        ...inferSelectors(element, root, excludeRandomClasses),
-        ...stableAncestors,
-      ])
-    ),
-    getSelectorPreference
-  );
-}
-
-/**
- * Generate some CSS selector variants for an element.
- */
-export function inferSelectors(
-  element: HTMLElement,
-  root?: Element,
-  excludeRandomClasses?: boolean
-): string[] {
-  const makeSelector = (allowed?: Array<keyof typeof CssSelectorType>) => {
-    try {
-      return safeCssSelector(element, {
-        selectors: allowed,
-        root,
-        excludeRandomClasses,
-      });
-    } catch (error) {
-      console.warn("Selector inference failed", {
-        element,
-        allowed,
-        root,
-        error,
-      });
-    }
-  };
-
-  return sortBy(
-    compact(
-      uniq([
-        makeSelector(["id", "class", "tag", "attribute", "nthchild"]),
-        makeSelector(["tag", "class", "attribute", "nthchild"]),
-        makeSelector(["id", "tag", "attribute", "nthchild"]),
-        makeSelector(["id", "tag", "attribute"]),
-        makeSelector(),
-      ])
-    ),
-    getSelectorPreference
-  );
-}
-
-/**
- * Returns true if selector uniquely identifies an element on the page
- */
-function doesSelectOneElement(selector: string): boolean {
-  return $safeFind(selector).length === 1;
-}
-
-export function getCommonAncestor(...args: Node[]): Node {
-  if (args.length === 1) {
-    return args[0].parentNode;
-  }
-
-  const [node, ...otherNodes] = args;
-
-  let currentNode: Node | null = node;
-
-  while (currentNode) {
-    // eslint-disable-next-line @typescript-eslint/no-loop-func -- The function is used immediately
-    if (otherNodes.every((x) => currentNode.contains(x))) {
-      return currentNode;
-    }
-
-    currentNode = currentNode?.parentNode;
-  }
-
-  return null;
-}
-
-export function findContainerForElement(element: HTMLElement): {
-  container: HTMLElement;
-  selectors: string[];
-} {
-  let container = element;
-  let level = 0;
-
-  if (BUTTON_TAGS.includes(element.tagName.toLowerCase())) {
-    container = element.parentElement;
-    level++;
-  }
-
-  /**
-   * If the direct parent is a list item or column, that's most li
-   */
-  if (MENU_TAGS.includes(container.parentElement.tagName.toLowerCase())) {
-    container = container.parentElement;
-    level++;
-  }
-
-  const extra: string[] = [];
-
-  if (container !== element) {
-    const descendent = level === 1 ? ">" : "> * >";
-
-    if (element.tagName === "INPUT") {
-      extra.push(
-        `${container.tagName.toLowerCase()}:has(${descendent} input[value="${escapeDoubleQuotes(
-          element.getAttribute("value")
-        )}"])`
-      );
-    } else {
-      extra.push(
-        `${container.tagName.toLowerCase()}:has(${descendent} ${element.tagName.toLowerCase()}:contains("${escapeDoubleQuotes(
-          $(element).text().trim()
-        )}"))`
-      );
-    }
-  }
-
-  return {
-    container,
-    selectors: uniq([
-      ...extra.filter((selector) => doesSelectOneElement(selector)),
-      ...inferSelectors(container),
-    ]),
-  };
-}
-
-export function findContainer(elements: HTMLElement[]): {
-  container: HTMLElement;
-  selectors: string[];
-} {
-  if (elements.length > 1) {
-    const container = getCommonAncestor(...elements) as HTMLElement | null;
-    if (!container) {
-      throw new Error("Selected elements have no common ancestors");
-    }
-
-    return {
-      container,
-      selectors: inferSelectors(container),
-    };
-  }
-
-  return findContainerForElement(elements[0]);
-}
-
-/**
- * Find direct children of the container that contain each selected descendent element
- * @param $container the panel container
- * @param selected the selected descendent elements
- */
-function containerChildren(
-  $container: JQuery,
-  selected: HTMLElement[]
-): HTMLElement[] {
-  return selected.map((element) => {
-    const exactMatch = $container.children().filter(function () {
-      return this === element;
-    });
-
-    if (exactMatch.length > 0) {
-      return exactMatch.get(0);
-    }
-
-    const match = $container.children().has(element);
-
-    if (match.length === 0) {
-      throw new Error("element not found in container");
-    }
-
-    return match.get(0);
-  });
-}
-
 export function inferSinglePanelHTML(
   container: HTMLElement,
   selected: HTMLElement
@@ -878,4 +528,32 @@ export function inferButtonHTML(
   throw new BusinessError(
     `Did not find any button-like tags in container ${container.tagName}`
   );
+}
+
+/**
+ * Find direct children of the container that contain each selected descendent element
+ * @param $container the panel container
+ * @param selected the selected descendent elements
+ */
+function containerChildren(
+  $container: JQuery,
+  selected: HTMLElement[]
+): HTMLElement[] {
+  return selected.map((element) => {
+    const exactMatch = $container.children().filter(function () {
+      return this === element;
+    });
+
+    if (exactMatch.length > 0) {
+      return exactMatch.get(0);
+    }
+
+    const match = $container.children().has(element);
+
+    if (match.length === 0) {
+      throw new Error("element not found in container");
+    }
+
+    return match.get(0);
+  });
 }
